requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.7.0=pypi_0
anyio=4.6.2.post1=pypi_0
authlib=1.3.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
cryptography=43.0.3=pypi_0
expat=2.6.3=h313beb8_0
grpcio=1.67.1=pypi_0
grpcio-health-checking=1.67.1=pypi_0
grpcio-tools=1.67.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httpx=0.27.0=pypi_0
idna=3.10=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
ncurses=6.4=h313beb8_0
numpy=2.1.3=pypi_0
openssl=3.0.15=h80987f9_0
pandas=2.2.3=pypi_0
pip=24.2=py312hca03da5_0
protobuf=5.28.3=pypi_0
pycparser=2.22=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
python=3.12.4=h99e199e_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
setuptools=75.1.0=py312hca03da5_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h80987f9_0
tk=8.6.14=h6ba3021_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
validators=0.34.0=pypi_0
weaviate-client=4.9.3=pypi_0
wheel=0.44.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
